backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# GitHub Integration
PyGithub==2.5.0
gitpython==3.1.43

# AI/ML
langchain==0.3.9
langchain-openai==0.2.9
langchain-community==0.3.9
openai==1.57.2
tiktoken==0.8.0

# Vector Store
pgvector==0.3.5

# Redis
redis==5.2.0
hiredis==3.0.0

# Utilities
httpx==0.28.1
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
